initial changes
fixed supply, no block rewards
on-demand blocks (no need to build up PoW depth since it will rely on
dPoW for security)
command line configurable chain parameters

the last one is a pandora's box as it allows anybody to spawn a new
fork just by specifying on the ./bitcoind command line the port,
netmagic, timestamp, symbol, supply. It will deterministically
calculate the nonce required for the genesis block, along with the
genesisblock from that and presto! we have a new BU chain